Fixes an issue in ByteArrayPageCursor
Where calling next() would always go to the next page.
This breaks the typical interaction contract with a PageCursor
where it's expected that the first call to next() places it
on the page it was instantiated to start on.

The place which discovered this was BaseHighLimitRecordFormat
where it would set up a composite page cursor. The secondary
cursor would be opened as a linked cursor placed at the calculated
page and offset and when a call to next() would be made
right after instantiation it would instead go to the next page
and write there, which is wrong.