Fixes issue with high_limit format scanning records using scan cursors
If a record read was aborted early in the high_limit format
it didn't set the offset of the cursor to the end of the record,
which is its contract to do. Instead it just left it where ever
it was. This was fine for most cases, but the storage-level
cursors "scan" operations took advantage of a store access
method which used the record reading to advance the cursor offset
itself instead of calculating it for every read. Forgetting to
set the offset properly would affect the next read(s) in
quite random ways as it would try to start reading the next
record at an arbitrary place in the store, not at the start
of the record and so data it observed would be total garbage.