Improve isPredicate check
`isPredicate` is used when compiling expressions to decide if we need to
coerce to a boolean at runtime. This is needed for expressions such as `WHERE (a)-->(b)`.
However the check was missing all the runtime expressions such as `HasLabelsFromSlot`
et al, since the method lives in the community module, leading to unnecessary coercions
in runtime.