IndexEntryUpdate can represent both property index and token index updates (#7975)
IndexEntryUpdate can represent both index updates and scan store updates

Scan store updates are currently represented by `TokenIndexEntryUpdate` and index updates are represented by `IndexEntryUpdate`. With this PR we let `IndexEntryUpdate` represent both scan store and index updates.

The idea is that we only care about what type of update we are making when we create the update, during population or online update, and when we apply the update, using an `IndexUpdater`. Each implementation of the `IndexProvider` umbrella with `IndexAccessor` and `IndexPopulator` knows which type of updates it deals with and will cast to the correct type if it needs to. We need to do this because we are depending on the inner structure of the update class, specifically if it holds `Value[]` or `long[]`.

Most tests are specific to the ValueIndexEntryUpdate so they just use that type directly.