Fix possible multiple default database creation
Previously if the system graph has one database marked as default,
but the config setting has another database as default (only possible
in a clustered setup), creating a database with the same name as the
config default one would mark that as default as well. This results
in two databases marked as default in the system graph.

Co-authored-by: Nadja Müller <nadja.muller@neo4j.com>