Fix issue where readers could be empty despite there being online cores in the routing table (#8162)
* Fix issue where readers could be empty despite there being online cores.

Behaviour should be to fallback to followers and or leader if there are no online read replicas. Previously this check was done too late.

To make AddressCollector more readable a new abstraction, ClusterServerInfos, is introduced. This provides an snapshot of server infos and their OperatorStates. The interface helps readability in AddressCollector.

* Easier to read that leader is removed from followers