Fix bug with DropResults and updates
For queries like

```
CREATE (n) WITH n WHERE 1 < 0 RETURN 42
```

when we can statically know that the predicate
is false we were rewriting to DropResult
leading to no nodes being created. Whereas
for a query like

```
CREATE (n) WITH n WHERE $p RETURN 42
```

with `p->false` we are creating a node.

This PR removes `DropResult` altogether and replaces
it with either `Limit(0)` or `ExhaustiveLimit(0)`.