Fix memory leak in cached PipelinedExecutionPlans
Through PipelinedExecutionPlan
          .executionGraphDefinition
          .pipelines
          .head
          .headPlan
          .asInstanceOf[FusedHead]
          .operatorFuser

we were caching all sorts of things that should not be cached.
But most importantly this class had a TokenContext with was
an instance of PlanContext.

This is now fixed by splitting the building of a Seq[NewTemplate]
and the compilation of that sequence in an Operator into two different classes.

One dirty cast was replaced with another in the process.
If we want to remove that, we need to make NewTemplate available in the
physicalplanning module.