Pick the leader core correctly
Before it was possible to pick a non-leader core by picking a candidate that wouldn't get elected.
Ultimately this led to flakiness on the test because
triggering an election on him would do nothing relevant and we would end up finding
the right leader later on sitting comfortably in term 1.

Now with this change we tap into the leader listener which can't lead us into picking
the wrong core as leader.