Parallel merging when building index after population
After initial population of a native index each "shard" of the index entries
will be sorted and merged in parallel, this is already the case.

The added parallel merging in this commit is about the merging of these
shards that happens after that. Since the writing of the index entries to
the GBPTree is done by a single thread then it was also previously so that
that thread would, for every entry, compare the heads of all the shards,
pick the lowest one and write it. This scales poorly with a high number
of shards. For reference with 32 shards the overhead of this
merging-by-comparison was larger than that of writing entries.
Now instead the merging of the shards happens in parallel with the writing
of the entries and will merge 4-by-4 shards in a tree-like processing
tree-like structure, like so:

```
  (n): a shard containing index entries from the scan

  (1) (2) (3) (4) (5) (6) (7) (8) (9) (10)
    \  |   |  /     \  |   |  /   /   /
     (---A---)       (---B---)   /   /
            \           |       /   /
          (----------C-------------)
```

And the writer, which is run by the main thread will only pull from C
in this scenario above, which is a single sorted stream of entries.
This makes the wall clock time spent only dependent on the writing
of the index entries, not the merging of them.