Ensure leader is follower after election
This is a safety guard against flakiness.
After triggering the election the leader should have become a follower.
If he hasn't then there was something wrong and we are better off retrying the test.

Note, however, that it is possible that we have been context switched
between the election trigger and become the leader or candidate for a new term.
This is unlikely though, and given we retry it's fine.