Bump record format versions for planned changed in 4.3
HighLimit format has been a bit different than the others
with regards to making this bump where most format files
have been copied (except only the dynamic record format class
for some reason). Tried to minimize the amount of code copied
in this bump so that it aligns with the how the other formats
are usually bumped.