Introduce KernelVersion
Which will act as a unification of versioning of the kernel/storage
e.g. log entry versions and the version a each database is set to run
with. The main driver for this is the essential change where
DbmsRuntimeVersion dictates the version that the dbms runs with
as a whole and the problem where the dbms runtime version is
set on the system db and needs to propagate to each database
transactionally to work properly.

The other aspect is to simplify versioning and future bumping
of version, for e.g. storage/log changes.