Adds database "dbms runtime version" upgrade listener
This will act as a transactional barrier, automatically committed
per database before the first write transaction to it after
it notices a change in the dbms runtime version.

Rolling upgrade scenario:

- Upgrade one member one instance at a time, where the restarted
  members run on new Neo4j version jars, although while still
  staying on an older kernel version.
- All members have been upgraded
- The upgrade procedure gets called which sets a new
  dbms runtime version. This will however not affect any database
  at this point in time since it's set on the system db
- The first write transaction for a database will notice this
  version change when comparing its kernel version to the
  kernel version that the current dbms runtime version specifies.
  When it does a database upgrade transaction will be committed
  first and then the actual transaction.
- Transactions after this point will all use the new version.

This transaction event listener will only be registered on
database startup if it notices that its kernel version is
not what the latest dbms runtime version specifies as its
kernel version. This listener will also unregister after having
performed an upgrade transaction which brings it to the version
specified by the latest dbms runtime version, so the performance
hit of having this listener is only had in the time window when
upgrade is possible.