First accesses for cached properties
If there are multiple accesses to the same property in the same plan, we cannot know which one will read first.
This is especially important for Selections, if two predicates use the same property.
We would otherwise reorder predicates such that the one which had `knownToAccessStore==false`
almost certainly would come first, which defeats the purpose of reordering the predicates.