Add database layout to IndexProviderFactory dependencies
The new token indexes need the database layout.

Adding this dependency also happened to fix that EnterpriseDbmsExtension now inject
a databaseLayout that matches the injected database. Before the injected database
would be a newly created one, while the injected database layout would be the one
belonging to the default database.