Importer uses larger batches for relationship linking
Generally makes reading performance more consistent and helps
a great deal in backwards scanning where each reader does
something like this:

```
Store low ID [                                                ] high ID
jump to start of batch 1                               V
read batch 1                                           ------>
jump to start of batch 2                        V
read batch 2                                    ------>
...
```

The larger the batches during backwards reading the more sequential
reading becomes. It almost turns into a forward-scan.

It has typically been the case that the relationship backwards scan
takes 3-4x longer than the forwards scan. Especially if the
relationship store is large enough that the OS file system cache
cannot fit it or most of it. But with this change it looks like
the performance of the backwards scan is as good as that of the
forwards scan, even if most of the relationship store cannot
fit in the fs cache, i.e. even if it actually needs to read from disk.