Fixed concurrency bug during startup
While the StartupOperator was starting a database the SystemGraphDbmsOperator might
receive an update for the state of that database. If the two reconciliations get
triggered to run concurrently the two operators have different desired state for the
database. The reconciler will prioritise certain states which in this case is not
necessarily the correct one.

Refactored a few lines in StartupOperator that was unused/duplicate command or
potentially thread unsafe.