Use read write lock in Segmented raft log
Most calls to SegmentedRaftLog are synchronized in the RaftMachine. But there are also some get calls in RaftLogShipper which could become problematic. This removed synchronized from the SegmentedRaftLog and replaces is with a read-write lock to ensure we do not try to read state while we are modifying the log