Add tests for standalone + read replicas clusters
- Leveraging the existent tests because it's the simplest
way forward:  most scenarios testing read replicas should
be valid tests when done for cores + rr as well as standalone + rr.

- Making StandaloneMember a ClusterMember also because it
is the easiest way to leverage all the existent test
machinery (e.g., DataCreator.createSchema( cluster ), coreTx , etc.,).
Arguably a bit hacky and misleading, but it has the benefit of making
core clusters tests compatible with standalone tests without
much work (as seen in ReadReplicaReplicationIT).

- I will work (immediately) on a follow up commit to rename
the usages of core to primary to make it clear that we are starting
either a standalone or core member.