Cache log file first transaction start time in TransactionLogFileTimestampMapper
Default time-based transaction log pruning strategy tries to extract the time of the first transaction in the log file to be able to determine if it's safe to delete that particular file. That's not a big deal while you do not have that many transaction log files to inspect and you do not have continuous checkpointing configured. If both of those things are true we will waste tons of io resources for nothing. Any checkpoint will try to determine if they're anything to be pruned and will inspect all files, continuously re-reading them.
As a result time, it is actually required to perform checkpoint itself can be a small fraction of the time required to check log running condition, and all of that continuously.
This PR will try to avoid that by caching the last 10k of observed log version tx timings.