Designated seeder
Allows the user to define a "designated seeder" when creating a database on a running cluster. This means that a backup only has to be restored on one core and the cluster will handle the replication of that data. The designated needs to have store files for the database. If other instances already have store files for the database they will be deleted.

Co-authored-by: Aleksey Karasavov <aleksey.karasavov@neotechnology.com>