Stop standalone db on failure (instead of shutdown) (#9445)
* Stop standalone db on failure (instead of shutdown)

We should make sure that we can start a database again after
a failure to start the first time. By calling stop, we ensure
that the life support of a standalone database can start again
(e.g., when a user explicitly calls it). The manager itself never
calls shutdown on the EnterpriseDatabase. We believe this is safe
because a database is restartable, and the shutdown of the database is called
on the global life lifecycle.