Changed logging when match index not progressing
AppendEntries.Response is now sent over a channel that doesn't enforce
ordering. If these messages are handled out of order on the leader that
instance will see the match index not progressing.

If newMatchIndex is less then matchIndex we assume the AppendEntries.Response
has arrived out of order which itself causes no concern.

If newMatchIndex is equal to matchIndex we are not progressing. This might happen
during some mode transitions and is fine if it is transient.