Fixes an issue where unnecessary relationship groups were created
When creating a relationship for a dense node to then delete it
in the same transaction before commit would leave behind a trace
in the transaction state such that at commit a relationship group
for that type would be created for that node, even tho it would
be empty.

This would later became problematic when trying to delete
that node because none of its deleted relationships would be of that
type and so that group would not be noticed and therefore not deleted.
This would result in the inability the delete that node because
validation logic would think that the node still had relationships left