Revert Init RR status lastTxId
Revert #7851 / 07f90adb1237249fb63a7f6e32ee59727698df12
Filling lastAppliedRaftIndex with a txId is not correct. Those two have no connection.
We have to live with the fact that after a storecopy lastAppliedRaftIndex will show 0 on a RR.
This should be a big problem, since lastAppliedRaftIndex make no sense on a RR which is not part of a raft group.