Indexing service can now populate token indexes
The population itself is a small part of the change.

The largest part are changes triggered by changing of signature
of store scan API. The reason for that is the replacement of
EntityTokenUpdate with TokenIndexEntryUpdate. TokenIndexEntryUpdate
has a generic parameter, which would lead to adding yet another
generic parameter to store scan methods and also the information
which token index is being populated will need to be propagated
to places where TokenIndexEntryUpdate is created.
Therefore the visitors accepted by store scan are replaced by
TokenScanConsumer and PropertyScanConsumer which as a side effect
cleaned the scan API a lot.