Increase akka actor system log timeout in tests
Saw a test (ReadReplicaViewMessage#shouldReturnDatabaseStates) failing on this with a default timeout of 5s.
It was only on PR's, but it's a legit flakiness, that could happen on other tests as well,
so I changed all unit test creations of the actor system to increase the timeout to
our default value.

```
[WARN] [03/04/2021 16:51:07.921] [ForkJoinPool-1-worker-3] [EventStream(akka://ReadReplicaViewMessage)] Logger log1-Logging$DefaultLogger did not respond within Timeout(5000 milliseconds) to InitializeLogger(bus)
```
and

```
Caused by: akka.event.Logging$LoggerInitializationException: Logger log1-Logging$DefaultLogger did not respond with LoggerInitialized, sent instead [TIMEOUT]
	at akka.event.LoggingBus.addLogger(Logging.scala:212)
	at akka.event.LoggingBus.$anonfun$startDefaultLoggers$4(Logging.scala:125)
	at scala.util.Success.$anonfun$map$1(Try.scala:255)
	at scala.util.Success.map(Try.scala:213)
	at akka.event.LoggingBus.$anonfun$startDefaultLoggers$3(Logging.scala:124)
	... 82 more
```