Hanging catchup requests
There was a concurrency issue if a channel is closed in the time frame between it being released to the channel pool until a new response handler has been registered. The new handler wound not be notified that the channel was closed and instead wait for a response on a closed channel.